Add fragment overlay class for a list of TriggerPrimitives
In order to save lists of trigger primitives in trigger records, we
need a way to pack them into fragments. This commit provides the
overlay class representing an ad hoc serialization of a list of
trigger primitive objects. (It was decided not to use the
functionality of the serialization package to achieve this, as it
would make fragments containing DS objects look different to fragments
containing data from hardware, and would require offline to depend on
a library to read the MsgPack or JSON format used by the serialization
package).

This commit is just the overlay: code to create Fragment objects and
fill them with TP data is in trigger/FragmentMakers.hpp, along with
code to read back the TPs out of Fragments